# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BugShot API Documentation",
        "description": "OpenApi documentation using swagger for BugShot project.",
        "contact": {
            "email": "info@it-michel.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.test/api/v1"
        }
    ],
    "paths": {
...
        "/companies": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "All companies.",
                "operationId": "allCompanies",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "locale",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-projects",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-statuses",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-bugs",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-screenshots",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-attachments",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-comments",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-company-users",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-project-users",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-bug-users",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-company-image",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-project-image",
                        "in": "header",
                        "required": false
                    },
                    {
                        "name": "include-attachment-base64",
                        "in": "header",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Company"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Store one company.",
                "operationId": "storeCompany",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "designation"
                                ],
                                "properties": {
                                    "designation": {
                                        "description": "The company name",
                                        "type": "string"
                                    },
                                    "color_hex": {
                                        "description": "The hexcode of the color (optional)",
                                        "type": "string"
                                    },
                                    "base64": {
                                        "description": "The base64 string of the image belonging to the company (optional)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/companies/{company_id}/image": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Company image.",
                "operationId": "showCompanyImage",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "company_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Company/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Image"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/companies/{company_id}/users": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "All company users.",
                "operationId": "allCompaniesUsers",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "company_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Company/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompanyUserRole"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Company": {
                "properties": {
                    "id": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "user_id": {
                        "description": "The id of the user that created the object.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "designation": {
                        "description": "The company name.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "color_hex": {
                        "description": "The colorcode for the company.",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "The creation date.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The last date when the resource was changed.",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "The deletion date.",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CompanyUserRole": {
                "properties": {
                    "company_id": {
                        "description": "The id of the company.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "user_id": {
                        "description": "The id of the user.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "role_id": {
                        "description": "The id of the role.",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        []
    ]
}